feat: add contribution guidelines and reorganize project structure
