+++ test.ts
@@ -5,6 +5,7 @@ import { registerFrontmatterTests } from "./src/utils/frontmatter_test.ts";
 import { registerScribeTests } from "./src/utils/scribe_test.ts";
 import { registerThemeTests } from "./src/theme/theme_test.ts";
 import { registerServerTests } from "./src/utils/server_test.ts";
+import { registerPublicApiTests } from "./src/public_api_test.ts";
 import { registerBuildTests } from "./src/core/build_test.ts";
 import { registerPluginTests } from "./src/plugins/plugins_test.ts";
 import { registerCollectionTests } from "./src/core/collections_test.ts";
@@ -16,6 +17,7 @@ registerFrontmatterTests();
 registerScribeTests();
 registerThemeTests();
 registerServerTests();
+registerPublicApiTests();
 registerBuildTests();
 registerPluginTests();
 registerCollectionTests();
